Refactors our common methods to global_helpers module
Moves rbac related methods to global_helpers.
Moves account and username related methods to global_helpers.
Defines Blueprint in global_helpers and registers it.
Defines global hooks using before_app_request in global_helpers.
