+++ system_baseline/app.py
@@ -14,6 +14,7 @@ from kerlescan.metrics_registry import create_prometheus_registry_dir
 from system_baseline import db_config, app_config
 from system_baseline.views.v1 import section as v1_bp
 from system_baseline.internal_views.v1 import section as internal_v1_bp
+from system_baseline.global_helpers import global_helpers_bp
 from system_baseline.models import db
 
 
@@ -60,6 +61,7 @@ def create_connexion_app():
 
     flask_app.register_blueprint(v1_bp)
     flask_app.register_blueprint(internal_v1_bp)
+    flask_app.register_blueprint(global_helpers_bp)
     flask_app.register_error_handler(HTTPError, handle_http_error)
     return connexion_app
